+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `no-restricted-properties` to not disallow certain object properties
 - Add `line-comment-position` rule to enforce position of line comments
 - Add `lines-around-directive` rule to require newlines around directives
+- Add `ignoreStrings`, `ignoreRegExpLiterals`, and `ignoreTemplateLiterals` options to `max-len` rule
 
 ### Changed
 - Change Gitter badge to a matching flat icon